move `impl pallet_collective::Config for Runtime` to separate file
