+++ runtime/devnet/src/pallets/mod.rs
@@ -18,3 +18,6 @@ pub use scheduler::*;
 
 pub mod preimage;
 pub use preimage::*;
+
+pub mod collective;
+pub use collective::*;
